# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Viz API
  version: 0.1.0
  description: OpenAPI specification for the Viz API (chi-based)
servers:
  - url: http://{host}:{port}
    description: Local API
    variables:
      host:
        default: localhost
        description: Hostname or IP for the API server
      port:
        default: "7770"
        description: Port for the API server
  - url: /api
    description: Production API
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /system/config:
    get:
      summary: Get system configuration
      operationId: getSystemConfig
      security:
        - BearerAuth: [admin:read]
        - CookieAuth: []
      responses:
        "200":
          description: System configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VizConfig"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /system/status:
    get:
      summary: Get system and user onboarding status
      operationId: getSystemStatus
      security: [] # Public endpoint, no auth required to check initial status
      responses:
        "200":
          description: System status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatusResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API key for authentication. The key should be provided in the Authorization header as a Bearer token. This token is a hex-encoded string of random bytes, prefixed with 'IMAG_'. Bearer format is intentionally omitted to avoid JWT confusion."
      x-scopes:
        admin:read: Read admin resources
        admin:write: Write admin resources
        api_keys:create: Create API keys
        api_keys:list: List API keys
        api_keys:get: Get API key details
        api_keys:revoke: Revoke API keys
        api_keys:rotate: Rotate API keys
        api_keys:delete: Delete API keys
        auth:refresh: Refresh authentication tokens
        auth:revoke: Revoke authentication tokens
        collections:create: Create collections
        collections:read: Read collections
        collections:update: Update collections
        collections:delete: Delete collections
        collections:share: Share collections
        images:read: Read images
        images:upload: Upload images
        images:update: Update images
        images:delete: Delete images
        images:download: Download images
        jobs:read: Read jobs
        jobs:create: Create jobs
        jobs:update: Update jobs
        jobs:delete: Delete jobs
        users:create: Create users
        users:read: Read users
        users:update: Update users
        users:delete: Delete users
        user_settings:read: Read user settings
        user_settings:update: Update user settings
        downloads:create: Create downloads
        events:read: Read events
        "*": Grants all permissions (superuser)

    CookieAuth:
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
...
    SystemStatusResponse:
      type: object
      properties:
        initialized:
          type: boolean
          description: True if the system has been initialized (at least one superadmin exists and first_run_complete is true).
        user_onboarding_required:
          type: boolean
          description: True if the current user (if authenticated) still needs to complete onboarding.
        needs_superadmin:
          type: boolean
          description: True if the system requires initial superadmin setup (no superadmin exists).
        allow_manual_registration:
          type: boolean
          description: True if user registration is enabled, let's users register accounts themselves
      required:
        - initialized
        - user_onboarding_required
        - needs_superadmin
        - allow_manual_registration
...
    VizConfig:
      type: object
      properties:
        baseUrl:
          type: string
          description: Base URL of the application
        logging:
          $ref: "#/components/schemas/LoggingConfig"
        base_directory:
          type: string
          description: Base directory path
        upload:
          $ref: "#/components/schemas/UploadConfig"
        database:
          $ref: "#/components/schemas/DatabaseConfig"
        redis:
          $ref: "#/components/schemas/QueueConfig"
        libvips:
          $ref: "#/components/schemas/LibvipsConfig"
        cache:
          $ref: "#/components/schemas/CacheConfig"
        user_management:
          $ref: "#/components/schemas/UserManagementConfig"
        storage_metrics:
          $ref: "#/components/schemas/StorageMetricsConfig"
    LoggingConfig:
      type: object
      properties:
        level:
          type: string
          description: Logging level
    UploadConfig:
      type: object
      properties:
        location:
          type: string
          description: Upload location
    DatabaseConfig:
      type: object
      properties:
        location:
          type: string
          description: Database location/host
        user:
          type: string
          description: Database user
        password:
          type: string
          description: "Masked password"
        name:
          type: string
          description: Database name
        port:
          type: integer
          description: Database port
    QueueConfig:
      type: object
      properties:
        enabled:
          type: boolean
          description: Is queue enabled
        host:
          type: string
          description: Queue host
        port:
          type: integer
          description: Queue port
        username:
          type: string
          description: Queue username
        password:
          type: string
          description: "Masked password"
        db:
          type: integer
          description: Redis DB index
        use_tls:
          type: boolean
          description: Use TLS
        pool_size:
          type: integer
          description: Connection pool size
        dial_timeout_seconds:
          type: integer
          description: Dial timeout
        read_timeout_seconds:
          type: integer
          description: Read timeout
        write_timeout_seconds:
          type: integer
          description: Write timeout
    LibvipsConfig:
      type: object
      properties:
        match_system_logging:
          type: boolean
          description: Match system logging level
        cache_max_memory_mb:
          type: integer
          description: Cache max memory MB
        cache_max_files:
          type: integer
          description: Cache max files
        cache_max_operations:
          type: integer
          description: Cache max operations
        concurrency:
          type: integer
          description: Concurrency level
        vector_enabled:
          type: boolean
          description: Vector enabled
    CacheConfig:
      type: object
      properties:
        gc_enabled:
          type: boolean
          description: GC enabled
    UserManagementConfig:
      type: object
      properties:
        allow_manual_registration:
          type: boolean
          description: Allow manual registration
    StorageMetricsConfig:
      type: object
      properties:
        enabled:
          type: boolean
          description: Metrics enabled
        interval_seconds:
          type: integer
          description: Interval in seconds
...