# viz/src/lib/types/config.types.ts
import type { SystemStatusResponse } from "$lib/api/client.gen";

export interface VizConfig {
    environment: "dev" | "prod";
    version: string;
    theme?: string;
}

export interface Vizig {
    system: SystemStatusResponse;
};